requirements.txt
discord.py==2.3.2
httpx==0.27.0
aiofiles==24.1.0
flask==3.0.3
gunicorn==23.0.0
